Allow owned and borrowed key value iterators to create attribute sets
