--- opentelemetry-sdk/src/attributes/mod.rs
@@ -1,1 +1,0 @@
-mod set;
