+++ opentelemetry/src/attributes/set.rs
@@ -156,6 +156,22 @@ impl Hash for InternalAttributeSet {
     }
 }
 
+pub trait ToKeyValue {
+    fn to_key_value(self) -> KeyValue;
+}
+
+impl ToKeyValue for KeyValue {
+    fn to_key_value(self) -> KeyValue {
+        self
+    }
+}
+
+impl ToKeyValue for &KeyValue {
+    fn to_key_value(self) -> KeyValue {
+        self.clone()
+    }
+}
+
 /// A unique set of attributes that can be used as instrument identifiers.
 ///
 /// Cloning of an attribute set is cheap, as all clones share a reference to the underlying
@@ -166,13 +182,15 @@ impl Hash for InternalAttributeSet {
 #[derive(Clone, Debug, Hash, PartialEq, Eq)]
 pub struct AttributeSet(Arc<InternalAttributeSet>);
 
-impl<'a, I: IntoIterator<Item = &'a KeyValue>> From<I> for AttributeSet
+impl<KV, I> From<I> for AttributeSet
 where
+    KV: ToKeyValue,
+    I: IntoIterator<Item = KV>,
     <I as IntoIterator>::IntoIter: DoubleEndedIterator + ExactSizeIterator,
 {
     fn from(values: I) -> Self {
         AttributeSet(Arc::new(InternalAttributeSet::from(
-            values.into_iter().cloned(),
+            values.into_iter().map(ToKeyValue::to_key_value),
         )))
     }
 }
@@ -226,6 +244,26 @@ mod tests {
         let array = [KeyValue::new("key1", "value1"), KeyValue::new("key2", 3)];
 
         let set = AttributeSet::from(&array);
+        let mut kvs = set.iter().collect::<Vec<_>>();
+
+        assert_eq!(kvs.len(), 2, "Incorrect number of attributes");
+
+        kvs.sort_by(|kv1, kv2| kv1.0.cmp(kv2.0));
+        assert_eq!(kvs[0].0, &Key::from("key1"), "Unexpected first key");
+        assert_eq!(
+            kvs[0].1,
+            &Value::String("value1".into()),
+            "Unexpected first value"
+        );
+        assert_eq!(kvs[1].0, &Key::from("key2"), "Unexpected second key");
+        assert_eq!(kvs[1].1, &Value::I64(3), "Unexpected second value");
+    }
+
+    #[test]
+    fn can_create_attribute_set_from_owned_vec() {
+        let vec = vec![KeyValue::new("key1", "value1"), KeyValue::new("key2", 3)];
+
+        let set = AttributeSet::from(vec);
         let mut kvs = set.iter().collect::<Vec<_>>();
 
         assert_eq!(kvs.len(), 2, "Incorrect number of attributes");
